refactor: dashboard 페이지 리팩토링
