--- types/column.ts
@@ -1,7 +1,0 @@
-export interface Column {
-  id: number
-  title: string
-  teamId: string
-  createdAt: string
-  updatedAt: string
-}
